220-check-python: prohibit python *.egg-info and *-nspkg.pth files with unknown version
imz@: The regex is made more general so that it matches even not yet
existing Python versions. Being ready to detect bad files in future is
better than silently ignoring them in future.

(imz@: In shell, "local" would hide errors which happened inside the command
substitution done on the same line; therefore, it's better to write
"local x" and "x="$(...)"" on separate lines.)